webapp/requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
branca==0.7.0
bs4==0.0.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Django==5.0
django-crispy-forms==2.1
django-jazzmin==2.6.0
django-shortuuidfield==0.1.3
fiona==1.9.5
folium==0.15.1
geographiclib==2.0
geopandas==0.14.2
geopy==2.4.1
gunicorn==21.2.0
humanfriendly==9.2
idna==3.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.4
networkx==3.2.1
numpy==1.26.3
osmnx==1.8.1
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
pycryptodome==3.20.0
pyproj==3.6.1
pyreadline==2.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
setuptools==69.0.3
shapely==2.0.2
shortuuid==1.0.11
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
ulid-py==1.1.0
urllib3==1.26.6
whitenoise==6.6.0
xyzservices==2023.10.1
django-ckeditor==6.7.0